refactor: separate exec/dmg.go
